+++ test/lo_model_tests.jl
@@ -272,3 +272,29 @@ end
     @test isapprox(result["solution"]["z_var"][2,5], 1.0)
     @test isapprox(result["solution"]["z_var"][3,4], 1.0)
 end
+
+@testset "Test topology flow cuts" begin
+    function data_I()
+        num_nodes = 8
+        instance  = 1
+        file_path = joinpath(@__DIR__, "..", "examples/instances/$(num_nodes)_nodes/$(num_nodes)_$(instance).json")
+        data_dict = LOpt.parse_file(file_path)
+        augment_budget = 7
+        return data_dict, augment_budget
+    end
+
+    data_dict, augment_budget = data_I()
+
+    params = Dict{String, Any}(
+        "data_dict"           => data_dict,
+        "augment_budget"      => augment_budget,
+        "eigen_cuts_full"     => false,
+        "eigen_cuts_2minors"  => true,
+        "topology_flow_cuts"  => true,
+        "time_limit"          => 1.5,
+    )
+
+    result = LaplacianOpt.run_LOpt(params, glpk_optimizer)
+    @test result["termination_status"] == MOI.TIME_LIMIT
+    @test result["primal_status"] == MOI.FEASIBLE_POINT
+end